+++ Resources/config/services.yml
@@ -4,14 +4,18 @@ parameters:
     zikula_dizkus_module.twig.extension.class: Zikula\DizkusModule\Twig\TwigExtension
     zikula_dizkus_module.links.class: Zikula\DizkusModule\Container\LinkContainer
     
-    zikula_dizkus_module.forum_helper.class: Zikula\DizkusModule\Helper\ForumHelper
-    zikula_dizkus_module.topic_helper.class: Zikula\DizkusModule\Helper\TopicHelper
-    zikula_dizkus_module.post_helper.class: Zikula\DizkusModule\Helper\PostHelper
+#    zikula_dizkus_module.forum_helper.class: Zikula\DizkusModule\Helper\ForumHelper
+#    zikula_dizkus_module.topic_helper.class: Zikula\DizkusModule\Helper\TopicHelper
+#    zikula_dizkus_module.post_helper.class: Zikula\DizkusModule\Helper\PostHelper
 
     zikula_dizkus_module.forum_user_helper.class: Zikula\DizkusModule\Helper\ForumUserHelper
     zikula_dizkus_module.moderators_helper.class: Zikula\DizkusModule\Helper\ModeratorsHelper
     zikula_dizkus_module.favorites_helper.class: Zikula\DizkusModule\Helper\FavoritesHelper
     zikula_dizkus_module.count_helper.class: Zikula\DizkusModule\Helper\CountHelper
+
+    zikula_dizkus_module.forum_manager.class: Zikula\DizkusModule\Manager\ForumManager
+    zikula_dizkus_module.topic_manager.class: Zikula\DizkusModule\Manager\TopicManager
+    zikula_dizkus_module.post_manager.class: Zikula\DizkusModule\Manager\PostManager
 
 services:
     
@@ -44,24 +48,6 @@ services:
 
 # Helpers
         
-    zikula_dizkus_module.forum_helper:
-        class: %zikula_dizkus_module.forum_helper.class%
-        tags:
-            - { name: zikula.dizkus.forum_helper }
-        arguments: [@request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @translator.default]
-
-    zikula_dizkus_module.topic_helper:
-        class: %zikula_dizkus_module.topic_helper.class%
-        tags:
-            - { name: zikula.dizkus.topic_helper }
-        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable]
-
-    zikula_dizkus_module.post_helper:
-        class: %zikula_dizkus_module.post_helper.class%
-        tags:
-            - { name: zikula.dizkus.post_helper }
-        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable]
-
     zikula_dizkus_module.forum_user_helper:
         class: %zikula_dizkus_module.forum_user_helper.class%
         tags:
@@ -86,3 +72,22 @@ services:
             - { name: zikula.dizkus.moderators_helper }
         arguments: [@doctrine.orm.default_entity_manager]
  
+# Managers
+        
+    zikula_dizkus_module.forum_manager:
+        class: %zikula_dizkus_module.forum_manager.class%
+        tags:
+            - { name: zikula.dizkus.forum_manager }
+        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable, @zikula_permissions_module.api.permission]
+
+    zikula_dizkus_module.topic_manager:
+        class: %zikula_dizkus_module.topic_manager.class%
+        tags:
+            - { name: zikula.dizkus.topic_manager }
+        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable, @zikula_dizkus_module.forum_manager]
+
+    zikula_dizkus_module.post_manager:
+        class: %zikula_dizkus_module.post_manager.class%
+        tags:
+            - { name: zikula.dizkus.post_manager }
+        arguments: [@translator.default, @router, @request_stack, @doctrine.orm.default_entity_manager, @zikula_users_module.current_user, @zikula_dizkus_module.security, @zikula_extensions_module.api.variable]
