+++ Resources/views/User/lastPostBy.html.twig
@@ -1,7 +1,7 @@
 <span>
     {% if last_post != null %}
-        {#  |profilelinkbyuname      #}
-        {{ __f("Last post by %s", {'%s':last_post.poster.user.uname}) }}<br />
+{#   @todo    profile link  #}  
+        {{ __f("Last post by %s", {'%s':last_post.poster.user.uname }) }}<br />
         {{ last_post.post_time|date('Y-m-d H:i:s') }}
         <a class="tooltips" title="{{ __("View latest post") }}" href="{{ lastTopicUrl(last_post.topic) }}"><i class="fa fa-hand-o-right"></i> {{ __('view') }}</a>
     {% endif %}
